+++ R/cNORM.R
@@ -47,6 +47,8 @@
 #'   \item CDC (2012). National Health and Nutrition Examination Survey: Questionnaires, Datasets
 #'   and Related Documentation. available: https://wwwn.cdc.gov/nchs/nhanes/OtherNhanesData.aspx.
 #'   date of retrieval: 25/08/2018
+#'   \item Harrel, F. (2020). Hmisc: Harrell Miscellaneous (v. 4.4-1). available https://CRAN.R-project.org/package=Hmisc
+#'   (code for weighted ranking adapted from wtd.rank & wtd.table by courtesy of Frank Harrell)
 #'   \item Lenhard, A., Lenhard, W., Suggate, S. & Segerer, R. (2016). A continuous solution to
 #'   the norming problem. Assessment, Online first, 1-14. doi: 10.1177/1073191116656437
 #'   \item Lenhard, A., Lenhard, W., Segerer, R. & Suggate, S. (2015). Peabody Picture Vocabulary